refactor: moved allow notification page into sign-up
