+++ src/pages/sign-up/allowNotification/allow-notification.module.scss
