+++ src/pages/sign-up/allowNotification/index.tsx
@@ -1,7 +1,7 @@
 import { useSelector } from 'react-redux';
 import { CurrentIdentity } from 'src/core/api';
 import { UserType } from 'src/core/types';
-import { getIdentityMeta } from 'src/core/utils';
+import { getIdentityMeta, translate } from 'src/core/utils';
 import { Button } from 'src/modules/general/components/Button';
 import { FeaturedIcon } from 'src/modules/general/components/FeaturedIcon';
 import { IconDropDown } from 'src/modules/general/components/iconDropDown';
@@ -9,9 +9,8 @@ import { RootState } from 'src/store';
 
 import css from './allow-notification.module.scss';
 import { useAllowNotification } from './useAllowNotification';
-export const AllowNotification = () => {
-  const type = localStorage.getItem('registerFor');
 
+export const AllowNotification = () => {
   const { onAllowNotification, items, onSkip } = useAllowNotification();
   const identities = useSelector<RootState, CurrentIdentity[]>(state => {
     return state.identity.entities;
@@ -19,6 +18,7 @@ export const AllowNotification = () => {
   const primary = identities.find(i => i.primary);
   const { name, username, profileImage } = getIdentityMeta(primary);
   const accounts = [{ id: '1', type: 'users' as UserType, name, username, img: profileImage }];
+  const type = localStorage.getItem('registerFor');
 
   return (
     <div className="flex flex-col h-screen justify-between items-center">
@@ -29,22 +29,20 @@ export const AllowNotification = () => {
       </div>
       <div className="flex flex-col h-screen justify-between">
         <div className="mt-20 text-center">
-          <h1 className={css.title}>Stay up to date</h1>
-          <h2 className={css.subTitle}>Turn on notifications to get updates</h2>
+          <h1 className={css.title}>{translate('sign-up-notification-header')}</h1>
+          <h2 className={css.subTitle}>{translate('sign-up-notification-subheader')}</h2>
         </div>
         <div className="flex flex-col justify-center items-center p-4">
           <FeaturedIcon iconName="bell-01" className="mb-5" />
-          <p className={css.description}>
-            Get alerts for new messages, job updates, and the progress of your current tasks.
-          </p>
+          <p className={css.description}>{translate('sign-up-notification-desc')}</p>
         </div>
         <div className=" w-full p-4">
           <Button color="primary" block onClick={onAllowNotification}>
-            Allow
+            {translate('sign-up-notification-allow')}
           </Button>
           <div className="mt-2">
             <Button color="secondary" variant="outlined" block onClick={onSkip}>
-              Skip
+              {translate('sign-up-notification-skip')}
             </Button>
           </div>
         </div>
